move decode_strings
