+++ tests/conftest.py
@@ -9,6 +9,7 @@ import viv_utils
 import floss.main as floss_main
 import floss.stackstrings as stackstrings
 import floss.tightstrings as tightstrings
+import floss.string_decoder
 from floss.const import DEFAULT_MIN_LENGTH
 from floss.identify import (
     get_function_fvas,
@@ -25,7 +26,9 @@ def extract_strings(vw):
     """
     top_functions, decoding_function_features = identify_decoding_functions(vw)
 
-    for s in floss_main.decode_strings(vw, get_function_fvas(top_functions), DEFAULT_MIN_LENGTH, disable_progress=True):
+    for s in floss.string_decoder.decode_strings(
+        vw, get_function_fvas(top_functions), DEFAULT_MIN_LENGTH, disable_progress=True
+    ):
         yield s.string
 
     no_tightloop_functions = get_functions_without_tightloops(decoding_function_features)
